new command "projects"
This new command shows a complete lists of projects.  It differs from
"list" in that it doesn't indicate which projects are currently installed
(making it much faster than "list"), in not having search capabilities,
and in showing only the project names (meaning that the output usually fits
on the screen).

Use this command in the the short help for newbies instad of list.

darcs-hash:20081029152746-e47ce-1e4ec4ab4039e8daa9ec174bea3deaf0a9d1617f.gz