Only look in directories of expanded glob for .asd files
When register_other_asd is handed a glob, it uses all the expanded
paths whether or not they are files.  The only reason this actually
works is because register_single_asd checks that the paths it is handed
are actually files.

darcs-hash:20081209151244-6c274-c6e76fb026931445e72329cf6f495eed70fd8a3a.gz